Add menu using Phaser
