+++ client/src/index.html
@@ -1,96 +1,41 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
 
-    <meta name="theme-color" content="#000000" />
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport"
+    content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover" />
+  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
 
-    <link rel="manifest" href="./manifest.json" />
+  <meta name="theme-color" content="#000000" />
 
-    <meta name="mobile-web-app-capable" content="yes" />
-    <meta name="apple-mobile-web-app-capable" content="yes" />
-    <meta
-      name="application-name"
-      content="<%= htmlWebpackPlugin.options.gameName %>"
-    />
-    <meta
-      name="apple-mobile-web-app-title"
-      content="<%= htmlWebpackPlugin.options.gameName %>"
-    />
-    <meta name="msapplication-starturl" content="./" />
+  <link rel="manifest" href="./manifest.json" />
 
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="192x192"
-      href="./icons/icons-192.png"
-    />
-    <link
-      rel="apple-touch-icon"
-      type="image/png"
-      sizes="192x192"
-      href="./icons/icons-192.png"
-    />
+  <meta name="mobile-web-app-capable" content="yes" />
+  <meta name="apple-mobile-web-app-capable" content="yes" />
+  <meta name="application-name" content="<%= htmlWebpackPlugin.options.gameName %>" />
+  <meta name="apple-mobile-web-app-title" content="<%= htmlWebpackPlugin.options.gameName %>" />
+  <meta name="msapplication-starturl" content="./" />
 
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="512x512"
-      href="./icons/icons-512.png"
-    />
-    <link
-      rel="apple-touch-icon"
-      type="image/png"
-      sizes="512x512"
-      href="./icons/icons-512.png"
-    />
+  <link rel="icon" type="image/png" sizes="192x192" href="./icons/icons-192.png" />
+  <link rel="apple-touch-icon" type="image/png" sizes="192x192" href="./icons/icons-192.png" />
+  <link rel="icon" type="image/png" sizes="512x512" href="./icons/icons-512.png" />
+  <link rel="apple-touch-icon" type="image/png" sizes="512x512" href="./icons/icons-512.png" />
 
-    <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon" />
+  <style>
+    html,
+    body {
+      height: 100%;
+      margin: 0;
+      background-color: #000000;
+    }
+  </style>
 
-    <title><%= htmlWebpackPlugin.options.gameName %></title>
-    <noscript
-      >Please enable javascript to continue using this application.</noscript
-    >
+  <noscript>Please enable javascript to continue using this application.</noscript>
+</head>
 
-    <!-- installs the serviceWorker -->
-    <!-- <script>
-      if ('serviceWorker' in navigator) {
-        window.addEventListener('load', () => {
-          navigator.serviceWorker.register('./sw.js')
-        })
-      }
-    </script> -->
-  </head>
+<body>
+  <div id="phaser-game" class="phaser-game"></div>
+</body>
 
-  <body>
-    <div class="center vertical">
-      <img
-        src="play/assets/img/logo.png"
-        onerror="this.onerror=null; this.src='assets/img/logo.png'"
-        alt="logo"
-        class="logo"
-        height="200"
-      />
-      <a href="/play">
-        <button>Play Game</button>
-      </a>
-      <br />
-      <button id="btnPlay" disabled>Host Game</button>
-      <div class="border disabled">
-        <h2>Dev Menu</h2>
-        <div>
-          <input value="12345" placeholder="Game ID" id="inputID" />
-          <div class="buttonContainer">
-            <button id="btnHost" disabled>Host ID</button>
-            <button id="btnJoin" disabled>Join ID</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </body>
 </html>
