--- client/src/play/index.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport"
-    content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover" />
-  <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-
-  <meta name="theme-color" content="#000000" />
-
-  <link rel="manifest" href="./manifest.json" />
-
-  <meta name="mobile-web-app-capable" content="yes" />
-  <meta name="apple-mobile-web-app-capable" content="yes" />
-  <meta name="application-name" content="<%= htmlWebpackPlugin.options.gameName %>" />
-  <meta name="apple-mobile-web-app-title" content="<%= htmlWebpackPlugin.options.gameName %>" />
-  <meta name="msapplication-starturl" content="./" />
-
-  <link rel="icon" type="image/png" sizes="192x192" href="./icons/icons-192.png" />
-  <link rel="apple-touch-icon" type="image/png" sizes="192x192" href="./icons/icons-192.png" />
-
-  <link rel="icon" type="image/png" sizes="512x512" href="./icons/icons-512.png" />
-  <link rel="apple-touch-icon" type="image/png" sizes="512x512" href="./icons/icons-512.png" />
-
-  <style>
-    html,
-    body {
-      height: 100%;
-      margin: 0;
-      background-color: #000000;
-    }
-  </style>
-
-  <noscript>Please enable javascript to continue using this application.</noscript>
-
-  <!-- installs the serviceWorker -->
-  <!-- <script>
-      if ('serviceWorker' in navigator) {
-        window.addEventListener('load', () => {
-          navigator.serviceWorker.register('./sw.js')
-        })
-      }
-    </script> -->
-</head>
-
-<body>
-  <a id="joinLink">
-    <h2 class="joinText" id="joinText"></h2>
-  </a>
-  <div id="phaser-game" class="phaser-game"></div>
-</body>
-
-</html>
