+++ client/src/scenes/mainScene.ts
@@ -1,28 +1,10 @@
 import { Direction } from "../../../common/types/direction";
+import { AnimationActor, AnimationSuffix } from "../types/animation";
 import { WorldState } from "../../../common/types/state";
 import { MultiplayerGame } from "../game/multiplayerGame";
 import { World } from "../objects/world";
+import { loadGameAssets } from "./load";
 import Sprite = Phaser.GameObjects.Sprite;
-
-const IDLE_FRAMERATE = 1;
-const RUN_FRAMERATE = 10;
-const ATTACK_FRAMERATE = 10;
-const DEATH_FRAMERATE = 10;
-const ZOMBIE_WALK_FRAMERATE = 8;
-const ZOMBIE_RUN_FRAMERATE = 8;
-
-export enum AnimationActor {
-  Player = "player",
-  Zombie = "zombie",
-}
-
-export enum AnimationSuffix {
-  Idle = "idle",
-  Run = "run",
-  Walk = "walk",
-  Attack = "atk",
-  Die = "die",
-}
 
 export function playAnimation(
   sprite: Sprite,
@@ -56,7 +38,6 @@ export default class MainScene extends Phaser.Scene {
     this.gameMaster = this.game.gameMaster;
     this.world = new World(this, this.game.gameMaster);
 
-    this.createAnimations();
     this.add.tileSprite(0, 0, 7680, 4320, "tiles").setDepth(-9999);
   }
 
@@ -69,129 +50,6 @@ export default class MainScene extends Phaser.Scene {
   }
 
   preload() {
-    this.load.image("tiles", "assets/Floor.png");
-    this.load.image("bullet", "assets/strip.png");
-    this.load.spritesheet("player", "assets/Player/rifle_map.png", {
-      frameWidth: 512,
-      frameHeight: 512,
-    });
-    this.load.spritesheet("zombie", "assets/Mobs/zombie_map.png", {
-      frameWidth: 512,
-      frameHeight: 512,
-    });
-  }
-
-  createAnimations() {
-    const directions = [
-      Direction.Down,
-      Direction.DownRight,
-      Direction.Right,
-      Direction.UpRight,
-      Direction.Up,
-      Direction.UpLeft,
-      Direction.Left,
-      Direction.DownLeft,
-    ];
-
-    directions.forEach((direction, index) => {
-      this.createAnimation(
-        AnimationActor.Player,
-        direction,
-        AnimationSuffix.Idle,
-        index * 8,
-        index * 8 + 1
-      );
-      this.createAnimation(
-        AnimationActor.Player,
-        direction,
-        AnimationSuffix.Run,
-        index * 8 + 2,
-        index * 8 + 7
-      );
-
-      this.createAnimation(
-        AnimationActor.Zombie,
-        direction,
-        AnimationSuffix.Walk,
-        index * 16,
-        index * 16 + 3,
-        ZOMBIE_RUN_FRAMERATE
-      );
-      this.createAnimation(
-        AnimationActor.Zombie,
-        direction,
-        AnimationSuffix.Run,
-        index * 16 + 4,
-        index * 16 + 7,
-        ZOMBIE_RUN_FRAMERATE
-      );
-      this.createAnimation(
-        AnimationActor.Zombie,
-        direction,
-        AnimationSuffix.Attack,
-        index * 16 + 8,
-        index * 16 + 11,
-        ZOMBIE_WALK_FRAMERATE
-      );
-    });
-
-    const directionsDie = [
-      Direction.Left,
-      Direction.DownLeft,
-      Direction.Down,
-      Direction.DownRight,
-      Direction.Right,
-      Direction.UpRight,
-      Direction.Up,
-      Direction.UpLeft,
-    ];
-
-    directionsDie.forEach((direction, index) => {
-      this.createAnimation(
-        AnimationActor.Zombie,
-        direction,
-        AnimationSuffix.Die,
-        index * 16 + 12,
-        index * 16 + 15,
-        DEATH_FRAMERATE
-      );
-    });
-  }
-
-  private createAnimation(
-    actor: AnimationActor,
-    direction: Direction,
-    suffix: AnimationSuffix,
-    startFrame: number,
-    endFrame: number,
-    frameRate?: number
-  ) {
-    let frameRateToUse = frameRate;
-    if (!frameRateToUse) {
-      switch (suffix) {
-        case AnimationSuffix.Idle:
-          frameRateToUse = IDLE_FRAMERATE;
-          break;
-        case AnimationSuffix.Run:
-          frameRateToUse = RUN_FRAMERATE;
-          break;
-        case AnimationSuffix.Attack:
-          frameRateToUse = ATTACK_FRAMERATE;
-          break;
-        case AnimationSuffix.Die:
-          frameRateToUse = DEATH_FRAMERATE;
-          break;
-      }
-    }
-
-    this.anims.create({
-      key: `${actor}-${direction}-${suffix}`,
-      frames: this.anims.generateFrameNumbers(actor, {
-        start: startFrame,
-        end: endFrame,
-      }),
-      frameRate: frameRateToUse,
-      repeat: suffix == AnimationSuffix.Die ? 0 : -1,
-    });
+    loadGameAssets(this);
   }
 }
