+++ client/src/types/animation.ts
@@ -1,0 +1,12 @@
+export enum AnimationActor {
+  Player = "player",
+  Zombie = "zombie",
+}
+
+export enum AnimationSuffix {
+  Idle = "idle",
+  Run = "run",
+  Walk = "walk",
+  Attack = "atk",
+  Die = "die",
+}
