+++ client/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2022",
-    "module": "es6",
+    "module": "es2022",
     "moduleResolution": "node",
     "allowJs": true,
     "sourceMap": true,
