+++ client/webpack/webpack.prod.cjs
@@ -38,8 +38,8 @@ const prod = {
     }),
     new CopyWebpackPlugin({
       patterns: [
-        { from: "src/assets", to: "play/assets" },
-        { from: "pwa", to: "play" },
+        { from: "src/assets", to: "assets" },
+        { from: "pwa", to: "" },
         { from: "src/favicon.ico", to: "" },
       ],
     }),
